move out supplier types from dir to file
